eureka-api-business-service-python/requirements.txt
aes-everywhere==1.2.10
aiofiles==0.7.0
anyio==3.3.0
appdirs==1.4.4
asgi-lifespan==1.0.1
asgiref==3.4.1
attrs==21.2.0
autopep8==1.5.7
black==21.7b0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.3
click==8.0.1
coverage==5.5
dependency-injector==4.41.0
Faker==8.11.0
fastapi==0.67.0
fastapi-versioning==0.9.1
firebase-admin==5.0.1
gcloud==0.17.0
google-api-core==1.34.0
google-api-python-client==2.93.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-appengine-logging==1.3.0
google-cloud-audit-log==0.2.5
google-cloud-core==2.3.2
google-cloud-firestore==2.2.0
google-cloud-logging==3.5.0
google-cloud-storage==1.41.1
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.58.0
greenlet==2.0.2
grpc-google-iam-v1==0.12.6
grpcio==1.56.0
grpcio-reflection==1.48.2
grpcio-status==1.48.2
grpcio-tools==1.48.2
h11==0.12.0
httpcore==0.13.6
httplib2==0.19.1
httptools==0.2.0
httpx==0.18.2
humanize==4.4.0
idna==3.2
importlib-metadata==6.0.0
iniconfig==1.1.1
jws==0.1.3
msgpack==1.0.2
mypy-extensions==0.4.3
oauth2client==3.0.0
packaging==21.0
pathspec==0.9.0
pluggy==0.13.1
proto-plus==1.22.3
protobuf==3.20.3
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodomex==3.10.1
pydantic==1.8.2
pyparsing==2.4.7
psycopg2-binary==2.9.4
pytest==6.2.4
pytest-asyncio==0.15.1
pytest-cov==2.12.1
python-dateutil==2.8.2
python-dotenv==0.19.0
python-jwt==2.0.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.7.6
requests==2.28.2
requests-toolbelt==0.7.0
rfc3986==1.5.0
rsa==4.7.2
sentry-sdk==1.4.3
sh==1.14.2
six==1.16.0
slack-bolt==1.16.1
slack-sdk==3.19.5
sniffio==1.2.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
sse-starlette==0.9.0
starlette==0.14.2
text-unidecode==1.3
toml==0.10.2
tomli==1.2.0
typed-ast==1.5.4
typing_extensions==4.7.1
uritemplate==3.0.1
urllib3==1.26.6
uvicorn==0.14.0
uvloop==0.17.0
versioneer==0.20
watchgod==0.7
websockets==9.1
zipp==3.11.0
